[nrf fromtree] net: sockets: Split native IP socket implementation from syscalls
Native IP socket implementation need only be build if native IP stack is
enabled. Therefore, split the native IP sockets from the common socket
syscalls shared across all socket implementations.

(cherry picked from commit 404e9c7b88629cdaf852c7c8d4d725c8f0c7117b)
